+++ service/status/status.go
@@ -50,7 +50,7 @@ type service struct {
 }
 
 func (s *service) Send(ctx context.Context, user *core.User, req *core.StatusInput) error {
-	if s.disabled || req.Build.Trigger == core.TriggerCron {
+	if s.disabled || req.Build.Event == core.EventCron {
 		return nil
 	}