+++ trigger/cron/cron_test.go
@@ -377,7 +377,7 @@ var (
 		Number:       dummyRepo.Counter,
 		RepoID:       dummyRepo.ID,
 		Status:       core.StatusPending,
-		Event:        core.EventPush,
+		Event:        core.EventCron,
 		Link:         "https://github.com/octocat/Hello-World/commit/7fd1a60b01f91b314f59955a4e4d4e80d8edf11d",
 		Timestamp:    1299283200,
 		Message:      "first commit",
@@ -446,7 +446,7 @@ var (
 	}
 
 	dummyHook = &core.Hook{
-		Event:        core.EventPush,
+		Event:        core.EventCron,
 		Link:         "https://github.com/octocat/Hello-World/commit/7fd1a60b01f91b314f59955a4e4d4e80d8edf11d",
 		Timestamp:    1299283200,
 		Message:      "first commit",